Crate lurk_metrics

Metrics infrastructure

This module hooks up the metrics facace to a thread-local metrics sink, that in turn is drained by a global recorder on a fixed cadence.

This crate is inspired by AWSLabs’ mountpoint-s3 (at v0.3.0)

Structs

MetricsSink

A global metrics sink that keeps a list of thread-local sinks to aggregate from

MetricsSinkHandle

Constants

METRICS_TARGET_NAME